增加有限域求逆和带余除法 内置数据类型由ulong改为长字节

+++ Cider.Compute/GFMatrix.cs
@@ -16,7 +16,7 @@ namespace Cider.Math
 
         #region Construtor
 
-        public GFMatrix(ulong row, ulong col, uint bitLength)
+        public GFMatrix(ulong row, ulong col, long bitLength)
         {
             Row = row;
             Column = col;
@@ -33,7 +33,7 @@ namespace Cider.Math
 
         public ulong Column { get; }
 
-        public uint BitLength { get; }
+        public long BitLength { get; }
 
         #endregion
